add basic styling
