+++ services/web/src/website/RegistrationFormPage/RegistrationFormPage.tsx
@@ -1,0 +1,12 @@
+import { Fragment } from 'react/jsx-runtime';
+import RegistrationForm from './RegistrationForm/RegistrationForm';
+import { Navigation } from '../HackathonPage/Navigation/Navigation';
+
+export const FormPage = () => {
+    return (
+        <Fragment>
+            <Navigation />
+            <RegistrationForm />
+        </Fragment>
+    );
+};
